Move phantom_server.js to ./lib/phantom.js and add package.json file.
Also, move caching class TotalRecall.js to ./lib

These changes are to prepares the code base to be set up as an npm
module (not quite ready yet...)

In the user-defined app.js file for an application, the following
command will be used like any other npm module:

    var phantom = require('phantom');

Until the phantom is registered with npm, this require statement can
be used as is if you do the following:

In the same directory as your app.js, create a "node_modules" subdirectory.

Then add a symlink called "phantom" pointing to the "phantom"
directory of your cloned phantom repo.