If query string params exists in a POST request, merge query string parameters with with body parameters.
This is what Google's app engine does (as well as others), so it's not
that weird. It's mostly a convenience in the real world.

If the developer makes a mistake and includes the same key in the
query string that already exists in the body, then the query string
parameter will override it.